Move code to prepare for e2e/restart transition
